cukinia-common: test syslog-ng config
Currently, a test verifies if syslog-ng service is the default logger
service, by checking if a log file `syslog.local` is present in
`/var/log/syslog-ng`. This test is not working because journalctl is the
default local logger, and syslog-ng send them over network to a distant
server.

This commit verifies that syslog-ng service is well configured to send
this log over network.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>